youtube_summarization_and_question_answering/requirements.txt
google-generativeai
python-dotenv
streamlit
youtube-transcript-api